--- a/hw10_program_optimization/Results.txt
+++ b/hw10_program_optimization/Results.txt
@@ -1,0 +1,15 @@
+from beginig
+    stats_optimization_test.go:46: time used: 781.869576ms / 300ms
+    stats_optimization_test.go:47: memory used: 308Mb / 30Mb
+
+after type users[100_000
+    stats_optimization_test.go:46: time used: 827.378995ms / 300ms
+    stats_optimization_test.go:47: memory used: 318Mb / 30Mb
+
+without regexp
+    stats_optimization_test.go:46: time used: 475.556759ms / 300ms
+    stats_optimization_test.go:47: memory used: 175Mb / 30Mb
+
+scanner
+    stats_optimization_test.go:46: time used: 624.75977ms / 300ms
+    stats_optimization_test.go:47: memory used: 93Mb / 30Mb