--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,11 +1,12 @@
 package hw10programoptimization
 
 import (
+	"bufio"
 	"encoding/json"
 	"fmt"
 	"io"
-	"regexp"
 	"strings"
+	"sync"
 )
 
 type User struct {
@@ -21,46 +22,45 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
+	users, errors := getUsers(r)
 	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
-		}
-
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+	for {
+		select {
+		case err := <-errors:
+			return nil, fmt.Errorf("get users error: %w", err)
+		case user, ok := <-users:
+			if !ok {
+				return result, nil
+			}
+			if strings.HasSuffix(strings.ToLower(user.Email), domain) {
+				result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]++
+			}
 		}
 	}
-	return result, nil
+}
+
+func getUsers(r io.Reader) (<-chan User, <-chan error) {
+	scanner := bufio.NewScanner(r)
+	users := make(chan User, 100)
+	errors := make(chan error)
+
+	wg := sync.WaitGroup{}
+	for scanner.Scan() {
+		wg.Add(1)
+		bytes := scanner.Bytes()
+		go func(bytes []byte) {
+			defer wg.Done()
+			var user User
+			if err := json.Unmarshal(bytes, &user); err != nil {
+				errors <- fmt.Errorf("unmarshal error: %w", err)
+			}
+			users <- user
+		}(append([]byte{}, bytes...))
+	}
+	go func() {
+		defer close(users)
+		defer close(errors)
+		wg.Wait()
+	}()
+	return users, errors
 }