… hw09_struct_validator/go.sum
… hw10_program_optimization/go.sum